chore: Refactor in preparation for remote deploys
